requirements.txt
streamlit
pandas
scikit-learn
SQLAlchemy
psycopg2-binary
python-dotenv
requests
streamlit-authenticator
plotly
numpy
sentry-sdk
wbdata